requirements.txt
python-telegram-bot==20.6
yt-dlp==2023.11.16
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
requests==2.31.0
python-multipart==0.0.6
